Implementando pablo_init

+++ pablo.c
@@ -86,7 +86,7 @@ int8_t basic_decoding_table[16] = {
 int8_t balanced_decoding_table[16] = {-127, -65, -50, -38, -28, -19, -10, -1, 0, 9, 18, 27, 37, 49, 64, 127};
 
 // pablo_q4_0 quantization radius
-int q4_0_radius = 0;
+uint8_t q4_0_radius = 0;
 
 // encoding-decoding table set to use
 enum _table_mode {BASIC_TABLE, BALANCED_TABLE};
@@ -139,7 +139,7 @@ void pablo_init(void) {
 
     // Initialize the rest of the data
     do_pablo = (strcmp(mode, "PBLO") == 0);
- 
+
     switch(table) {
 
         case BASIC_TABLE :
@@ -163,7 +163,7 @@ void pablo_init(void) {
         else 
             printf("PABLO:   performing SIMPLE (de)quantization\n");
 
-        printf("PABLO:   using radius %d\n", q4_0_radius);
+        printf("PABLO:   using radius %u\n", q4_0_radius);
 
         switch(table) {
             case BASIC_TABLE :
